LayoutTests/fast/selectors/first-child-update-with-renderer-expected.txt

:not(:first-child) should be worked for the first element beyond rendering

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById("target4")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("target5")).backgroundColor is "rgb(10, 11, 12)"
PASS getComputedStyle(document.getElementById("target6")).backgroundColor is "rgb(10, 11, 12)"
PASS getComputedStyle(document.getElementById("target6")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("target4")).backgroundColor is "rgb(10, 11, 12)"
PASS getComputedStyle(document.getElementById("target5")).backgroundColor is "rgb(10, 11, 12)"
PASS successfullyParsed is true

TEST COMPLETE
